modify compilation warning

+++ device/timer.c
@@ -3,8 +3,8 @@
 #include "print.h"
 
 #define IRQ0_FREQUENCY	   1
-#define INPUT_FREQUENCY	   1193180
-#define COUNTER0_VALUE	   INPUT_FREQUENCY / IRQ0_FREQUENCY
+#define INPUT_FREQUENCY	   1193180U
+#define COUNTER0_VALUE	   ((uint16_t)INPUT_FREQUENCY / (uint16_t)IRQ0_FREQUENCY)
 #define CONTRER0_PORT	   0x40
 #define COUNTER0_NO	   0
 #define COUNTER_MODE	   2
